Fix validation issues with the full_cr.yaml
The OpenAPIV3 schema is used to validate the YAML manifest, according
to it the cpu resource value must be a string or integer. For float
values it can be something like "0.2" (a string). This gets converted
to 200m internally. Setting the values to the internal representation
to avoid any confusion.

https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/#meaning-of-cpu

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>